create channel_merger.rs
